+++ Runtime/Scripts/ScriptableVariables/ScriptableVariable.cs
@@ -3,6 +3,7 @@ using Newtonsoft.Json;
 using Unity.Properties;
 using UnityEngine;
 using UnityEngine.Events;
+using Logger = Dev.Nicklaj.Butter.Helpers.Logger;
 
 namespace Dev.Nicklaj.Butter
 {
@@ -11,6 +12,7 @@ namespace Dev.Nicklaj.Butter
         [SerializeField] T _initialValue;
         [SerializeField] T _value;
 
+		public bool UseDebugLogs = false;
         public UnityAction<T> OnValueChanged = delegate { };
 
         [CreateProperty]
@@ -21,6 +23,8 @@ namespace Dev.Nicklaj.Butter
             {
                 if (this._value.Equals(value)) return;
                 this._value = value;
+				if(UseDebugLogs) 
+                    Logger.LogInfo($"Variable {this.name} has changed to {value}.");
                 OnValueChanged?.Invoke(value);
             }
         }
@@ -33,6 +37,8 @@ namespace Dev.Nicklaj.Butter
         [Obsolete("This should only be used when creating new scriptable variables, avoid calling this directly through code otherwise, it's here just for compatibility reasons.")]
         protected override void OnReset()
         {
+            if(UseDebugLogs) 
+                Logger.LogInfo($"Variable {this.name} has been reset to {_initialValue}.");
             OnValueChanged?.Invoke(_value = _initialValue);
         }
     }